+++ lib/elasticquery/filters/not.rb
@@ -1,4 +1,4 @@
-require_relative 'base'
+require_relative "base"
 
 module Elasticquery
   module Filters
@@ -9,7 +9,7 @@ module Elasticquery
       end
 
       def dup_with(*args)
-        raise StandardError, 'Cannot use Filters::Not twice'
+        raise StandardError, "Cannot use Filters::Not twice"
       end
 
       def to_hash
@@ -18,8 +18,8 @@ module Elasticquery
           filter = filters.last.dup_with condition
           if filter.valid?
             subquery = filter.to_hash
-            q = subquery[:query][:filtered][:filter][:and]
-            {query: {filtered: {filter: {and: {not: {filter: q}}}}}}
+            q = subquery[:query][:filtered][:filter][:and][0]
+            {query: {filtered: {filter: {and: [{not: {filter: q}}]}}}}
           else
             {}
           end