+++ test/test_helper.rb
@@ -1,5 +1,5 @@
-require 'minitest/test'
-require 'minitest/autorun'
+require "minitest/test"
+require "minitest/autorun"
 
 require "minitest/reporters"
 Minitest::Reporters.use! [Minitest::Reporters::DefaultReporter.new(:color => true)]